pyproject.toml
[tool.poetry]
name = "drone-swarms"
version = "0.2.0"
description = "A Python-based drone swarm simulator with real-time 3D visualization."
authors = ["Mew-72 <https://github.com/Mew-72>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.3"
matplotlib = "^3.10.0"
Pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"